+++ src/Fragments/App/Actions/ReorderFragments.php
@@ -3,15 +3,15 @@
 namespace Thinktomorrow\Chief\Fragments\App\Actions;
 
 use Thinktomorrow\Chief\Fragments\Events\FragmentsReordered;
-use Thinktomorrow\Chief\Shared\Helpers\SortModels;
+use Thinktomorrow\Chief\Shared\Concerns\Sortable\ReorderModels;
 
 class ReorderFragments
 {
-    private SortModels $sortModels;
+    private ReorderModels $reorderModels;
 
-    public function __construct(SortModels $sortModels)
+    public function __construct(ReorderModels $reorderModels)
     {
-        $this->sortModels = $sortModels;
+        $this->reorderModels = $reorderModels;
     }
 
     public function handle(string $contextId, array $indices, ?string $parentId = null): void
@@ -23,7 +23,7 @@ class ReorderFragments
         $where = 'context_id = "'.$contextId.'"';
         $where .= $parentId ? ' AND parent_id = "'.$parentId.'"' : '';
 
-        $this->sortModels->handle('context_fragment_tree', $indices, 'order', 'child_id', false, $where);
+        $this->reorderModels->handle('context_fragment_tree', $indices, 'order', 'child_id', false, $where);
 
         event(new FragmentsReordered($contextId));
     }